backend/requirements/common.txt
django~=3.0.0
djangorestframework
django-cors-headers
pillow
